Handle and persist theme color (#647)
## Fix knowing about when Themes themselves change

`window.hassConnection` is `undefined` when the `.js` is initially injected, so none of the registrations were working. This adds what is effectively a runloop wait (it is available 0ms later, it's really just an ordering thing) to make sure the connection exists.

Fixes #557.

## Fix not having the information anymore about the Themes

This now uses JavaScript to pull the underlying css variable for the particular theme colors, which should be a less fragile method to try and get it -- rather than parsing the actual CSS itself. We only care about the currently-active values, so we don't need to parse any other ones.

## Hackily observe when the current Theme changes

This is likely to break in the future and waits on home-assistant/frontend#4220 to correctly notify us when the change occurs.

## Inspect the Theme color and set an appropriate status bar color

Tries to set the right status bar text color based on the brightness of the given background color. I like the idea of trying to solve this for all cases, rather than relying on a Theme to customize it. Open to other ideas, though.

Fixes #645.