Reconnect WebView on a warm start if it failed to load (#653)
If the WebView failed to load in the past, and we've successfully connected to the API (which we now do on both cold and warm starts since #648), trigger a reload for the case that it failed to load entirely. Mostly just adds the case that a nil url is valid cause to reload.