Fix extension network requests having issues in beta (#778)
Feels like every time I touch a background URL session I fall into this same trap. From the [Apple docs](https://developer.apple.com/library/archive/documentation/General/Conceptual/ExtensibilityPG/ExtensionScenarios.html):

> Because only one process can use a background session at a time, you need to create a different background session for the containing app and each of its app extensions. (Each background session should have a unique identifier.) It’s recommended that your containing app only use a background session that was created by one of its extensions when the app is launched in the background to handle events for that extension. If you need to perform other network-related tasks in your containing app, create different URL sessions for them.

What this means is that we need to create different background URL sessions for extensions, and make the main app understand how to deal with the callbacks for these sessions. So this refactors a few ways:

1. We never refer to things just by the taskIdentifier, which is unique only per-session and not globally. Instead, we commingle the session identifier with the task identifier.

Thankfully, because I'm abusing the taskDescription on the tasks to store persistent information, we're not keying off anything for the task identifier except at runtime, so we don't need to worry about modifying any persistence to deal with this, which would have been a royal pain.

2. Whenever we're told to handle background events for a particular session, we always trust that it should exist and create it. This moves from having 1 background session to 1+N background sessions at any given time.

The secondary background sessions are never used for anything except handling incoming events for them. When they finish sending events, we kill them off so that we're not keeping control over the session with that identifier so the extensions can go back to using them, since there can only be one global background session per identifier.

Fixes #777.