Wrap webhook networking in background tasks (#781)
I have a few theories about #777:

1. The system isn't bothering to execute the background request because we've exhausted the background allotment that it wants for us. This is [somewhat documented (descriptively)](https://developer.apple.com/documentation/foundation/url_loading_system/downloading_files_in_the_background?language=objc):

> When the system resumes or relaunches your app, it uses a rate limiter to prevent abuse of background downloads. When your app starts a new download task while in the background, the task doesn't begin until the delay expires. The delay increases each time the system resumes or relaunches your app.

If this is what's happening, this fix won't do anything; the real fix will be to try to send ephemerally and drop to using a background request if that fails or times out.

I don't like this solution because it's gross! But, it'll be the next thing to try.

2. We're executing the actions but since we're allowing ourselves to be suspended right away the system doesn't consider it worth actually executing. Like, is this a heuristic maybe?

If this is the case, and I hope it is, then wrapping things in a background task will resolve this problem. It doesn't seem like this is it, but it's nice to do anyway, so let's give it a try.

----

Since this is adding a new failure mode to the register->re-integrate flow, adds another check to make sure we don't accidentally re-register when we don't need to.