Add update signaling to sensors; update Camera Sensor (#951)
Adds a 'dependencies' mechanism injected via the SensorProvider's request. Currently only one exists generically, and it's an entry point for an update signaler. This signaler object is cached and allocated when requested which allows SensorProviders to do arbitrary work and inform a handler when they're ready for an update.

The Camera Sensor uses this to observe 2 things:
1. The cameras connected to the device are changed (e.g. attached or detached)
2. The camera's is-on status changes, so it's transitioning from in-use to not-in-use