Remove Internal URL setting in Catalyst, bump build (#955)
Apple's documentation for the CaptiveNetwork API we use to query SSID information reads:

> This function returns NULL in Mac apps built with Mac Catalyst.

The header file for the same thing reads:

> On Mac Catalyst platform, to receive current Wi-Fi network information, an application must have "com.apple.developer.networking.wifi-info" entitlement and user's authorization to access location.

It seems the former is correct - even with the "access Wi-Fi info" entitlement specified, we cannot access Wi-Fi info on Catalyst, nor can I add the entitlement in the provisioning center. Hides the setting until I can figure out how to get to the information -- CoreWLAN on macOS has access to the data, but we can't access it directly since the contents of the framework are not available in Catalyst (but are, oddly, `import`-able).

Also adds the read-file/read-downloads entitlements since they are needed for weird downloaded-to-`~/Downloads` problems.