Fix slightly-blurry scale issues with WebView (#1043)
When the NSView's scaleFactor is 0.77 and we set the WebView to be 1.30 to offset it, we end up slightly scaling the text since it's not a perfect multiplication.

This eliminates the scale factor entirely when the scene is for the WebView, but leaves other scenes (e.g. settings, about) alone since we do want the automatic scaling in them.

Fixes #1041.