Add Android Application Record to NFC tag writing (#1149)
This is an external record which hints which application should be launched for the NFC tag. When uninstalled, this launches the app in the Play Store when read.

With this additional record, the payload size grows from 63 bytes to 115 bytes. Since this could potentially increase beyond a tag limit, the Android record will be dropped if the capacity could hold just the URI on its own.