Collapse microphones and cameras into singular sensors (#1320)
Breaking change. Fixes #1261.

Collapses all Camera and Microphone sensors from one-sensor-per-device to two camera and two microphone sensors:
- `binary_sensor.[device]_microphone_in_use` and `binary_sensor.[device]_camera_in_use` (value: true/false)
- `sensor.[device]_active_microphone` and `sensor.[device]_active_camera` (value: name or "Inactive")

This is unfortunately required because both the the `kCMIODevicePropertyDeviceUID`/`kAudioDevicePropertyDeviceUID` and the `kCMIOObjectPropertyName`/`kAudioObjectPropertyName` are not stable; the IDs change and names gain `#2`, `#3`, etc., suffixes. This is despite the documentation reading:

> A CMIODevice's UID is persistent across boots.

and

> An AudioDevice's UID is persistent across boots.

Were this the case, we wouldn't need to worry -- we unique on the device ID after all, and names can change over time were that part actually stable.