Fix crash deleting action in a new notification category (#1359)
Fixes #1256.

## Summary
When the category isn't persisted yet, we were still persisting the action into Realm. This means the action was deleted successfully, but the delete didn't cascade because the category wasn't yet persisted. The final result of this was that the category could refer to deleted objects, which crash when accessed.

## Screenshots
n/a

## Link to pull request in Documentation repository
n/a