Add backup timer for watch connectivity background task completion (#1454)
## Summary
Potentially fixes an issue where we do not cancel a watch connectivity request in time and get killed by the system.

## Any other notes
I cannot find any code paths that may be causing this issue, and I can't get it to happen in debugging (not that debugging with the watchOS simulator is, you know, usable -- wow it's so bad) but I can at least validate that we are calling it within the watchdog timeout from the system, which is 15.0s.

There's probably an issue somewhere in the path between using HA talking to Communicator talking to WatchConnectivity. I don't feel super confident in the Communicator code paths -- there's cases where e.g. failing to decide or something may not trigger handlers.

For a little bit I was distracted by the complication updating, but I already took that into account and made those requests not delayed by network requests (it fires passively, which doesn't wait for the response to return). So, I changed that up a little bit to be easier to read and reason about.