Fix crash on macOS 11.3 (#1500)
Fixes #1499.

## Summary
Fixes a crash caused by infinitely cursing into `setHa_scaleFactor(_:)` due to (of course) swizzling!

## Any other notes
Apple renamed (which is totally their prerogative, but it's also my prerogative to swizzle if they aren't going to make public API to do what I want 😈 ) the `scaleFactor` multiple for non-Mac optimized Catalyst builds, renaming it to `sceneToSceneViewScaleFactor`. Since we weren't checking if the swizzle was successful, the code path where we attempt to call into the swizzled variant calls itself infinitely until it gets killed.

Verified that the scale factor using this new property correctly doesn't make the webview super tiny, and hopefully made this less likely to break in the future.