Fix crash on iOS 14.5 due to invalid location data (#1504)
Fixes #1498.

## Summary
Ignores secondary location data (like vertical accuracy) which is NaN or infinite; the latter of which CLLocationManager is happily delivering at least in the iOS 14.5 betas.

## Any other notes
The problem with values like infinity is that they can't be represented in JSON, which means the system will raise an exception during JSON serialization. What's extra troublesome and ridiculous is that the JSONSerialization methods are marked `throws` -- as in, they throw Swift errors -- but the only error scenarios I can fathom (or can see in the [public implementation](https://github.com/apple/swift-corelibs-foundation/blob/99d6439016c008962d1b1cfc299dbbfd207c6454/Sources/Foundation/JSONSerialization.swift)) are ones that it raises exceptions for. Effectively, it never throws; it always raises an Objective-C exception.