Support time-sensitive notifications on iOS 15 (#1676)
## Summary
- Supports setting interruption-level via local push. Fixes #1658.
- Configures entitlement and notification setting. Fixes #1659.

## Link to pull request in Documentation repository
Documentation: home-assistant/companion.home-assistant#540

## Any other notes
Early beta days, I can't tell if the entitlement is all that we need, because I need the notification settings in the simulator which really doesn't like mutated-while-compiling entitlements. It's possible once we switch to Xcode 13 this may not need to be mutated at runtime, and the setting can go away and it'll all be happy.