Handle invalidly close to 0.0 latitude/longitude as invalid too (#1805)
Refs [a forum post](https://community.home-assistant.io/t/ios-15-location-issues/323498).

## Summary
iOS 15 appears to give us extremely small values for longitude like `9.368162246e-315`, which we should throw away as invalid.

## Any other notes
I reduced the logging we do for longitude/latitude, but we don't send it anywhere anymore (since only errors go up to Sentry on crash), so I'm adding it back in so I can see in future logs what's happening. Some of the logs from the user show the invalid value happening _multiple_ times and I just can't tell if we're butchering it from the raw, or if the raw coming in is bad. This should help.