Fix onboarding timeout when only internal URL is set and we lack SSID info (#1969)
Fixes #1918.

## Summary
During initial onboarding (or even after the fact), we don't have permission to read SSID values due to Location permission occurring later. When only an internal URL was set, we were falling back to homeassistant.local.:8123 which is possibly not valid, rather than using the one we slurped from discovery info.

## Any other notes
Manually entering a URL worked here because it sets the value as 'external URL' which didn't have this issue.