Decline opening a WebView for AirPlay to mirror instead (#1998)
The default behavior here is to open a second scene for the external display, but since you can't interact with it, it's not super useful.

We could maybe do something like:
- Offer to track the current page in the phone
- Add a setting to control which page to show, e.g. mirror or display

The upside of the second WebView scene is that it's at the full resolution of the device (e.g. at 1080p landscape, not whatever the phone is showing).