Fix infinite notification cycle when deleting server & servers resurrecting (#2190)
## Summary
A server change (e.g. active URL changing) triggers a save [which is
rejected due to being deleted] which causes a notification which causes
a server change which… ad infinitum. This prevents notifying that
connection info changed when the server save was blocked for some
reason.

This also moves the 'deleted server' identifier cache to be in
UserDefaults so extensions don't resurrect deleted servers, which seems
to be a recurring problem.

## Screenshots
<!-- If this is a user-facing change not in the frontend, please include
screenshots in light and dark mode. -->

## Link to pull request in Documentation repository
<!-- Pull requests that add, change or remove functionality must have a
corresponding pull request in the Companion App Documentation repository
(https://github.com/home-assistant/companion.home-assistant). Please add
the number of this pull request after the "#" -->
Documentation: home-assistant/companion.home-assistant#

## Any other notes
<!-- If there is any other information of note, like if this Pull
Request is part of a bigger change, please include it here. -->