simplify for loop (#3177)
<!-- Thank you for submitting a Pull Request and helping to improve Home
Assistant. Please complete the following sections to help the processing
and review of your changes. Please do not delete anything from this
template. -->

## Summary
Really small mr, I just came across it looking into the scripts widget.
I assume the for loop only returns when all sync operations are
finished, removing the need to manually check for it

The preview still seems to work fine when adding a widget